Change crlf line endings into lf
The convention when using git for platform-agnostic projects
is to check in LF line endings.

The Arduino IDE will write the file with line endings depending
on your platform: CRLF on Windows, LF on Linux and Mac.
https://github.com/arduino/Arduino/commit/3e50aee4cf56afdbd6490105d89eff9f2db447a7

For this reason, Windows users are encouraged to enable the
'core.autocrlf' option that makes sure CRLF line endings are
used locally, but LF line endings are used when committing.